+++ app/task/moph-erp.js
@@ -4,6 +4,8 @@ exports.erpAdminRequest = exports.updateAlive = exports.sendBedNo = exports.send
 const moment = require("moment");
 const moph_refer_1 = require("../middleware/moph-refer");
 const hismodel_1 = require("./../routes/his/hismodel");
+const os_1 = require("os");
+const fs = require('fs');
 const utils_1 = require("../middleware/utils");
 const packageJson = require('../../package.json');
 const dbConnection = require('../plugins/db');
@@ -195,6 +197,7 @@ exports.sendBedNo = sendBedNo;
 const updateAlive = async () => {
     const ipServer = (0, utils_1.getIP)();
     try {
+        const apiEnv = await detectRuntimeEnvironment();
         let data = {
             api_date: global.apiStartTime,
             server_date: moment().format('YYYY-MM-DD HH:mm:ss'),
@@ -203,7 +206,11 @@ const updateAlive = async () => {
             subversion: packageJson.subVersion || '',
             port: process.env.PORT || 0,
             ip: ipServer.ip,
-            his: hisProvider, ssl: process.env?.SSL_ENABLE || null,
+            host_type: apiEnv || 'host',
+            nodejs: process.version || '',
+            platform: (0, os_1.platform)() || '',
+            os_version: (0, os_1.release)() || '',
+            his: hisProvider, ssl: process.env?.SSL_ENABLE || 0,
         };
         const result = await (0, moph_refer_1.updateHISAlive)(data);
         const status = result.status == 200 || result.statusCode == 200 ? true : false;
@@ -211,12 +218,12 @@ const updateAlive = async () => {
             console.log(moment().format('HH:mm:ss'), '✅ Sent API Alive status result', result.status || '', result.statusCode || '', result?.message || '');
         }
         else {
-            console.log(moment().format('HH:mm:ss'), '❌ Sent API Alive status result', result.status || '', result.statusCode || '', result?.message || '');
+            console.error(moment().format('HH:mm:ss'), '❌ Sent API Alive status result', result.status || '', result.statusCode || '', result?.message || '');
         }
         return result;
     }
     catch (error) {
-        console.log(moment().format('HH:mm:ss'), '❌ Sent API Alive status error:', error?.status || error?.statusCode || '', error?.message || error || '');
+        console.error(moment().format('HH:mm:ss'), '❌ Sent API Alive status error:', error?.status || error?.statusCode || '', error?.message || error || '');
         return [];
     }
 };
@@ -325,4 +332,33 @@ function toSnakeCase(value) {
         return value;
     }
     return value.replace(/([a-z0-9])([A-Z])/g, '$1_$2').toLowerCase();
+}
+async function detectRuntimeEnvironment() {
+    let env = "host";
+    if (fs.existsSync("/.dockerenv")) {
+        env = "docker";
+    }
+    else {
+        try {
+            const cgroup = fs.readFileSync("/proc/1/cgroup", "utf8");
+            if (/docker|containerd/i.test(cgroup)) {
+                env = "docker";
+            }
+            if (/kubepods/i.test(cgroup)) {
+                env = "kubernetes";
+            }
+        }
+        catch { }
+    }
+    const r = (0, os_1.release)().toLowerCase();
+    try {
+        const version = fs.readFileSync("/proc/version", "utf8").toLowerCase();
+        if (r.includes("microsoft") ||
+            r.includes("wsl") ||
+            version.includes("microsoft")) {
+            env = "wsl";
+        }
+    }
+    catch { }
+    return env;
 }