+++ src/app.ts
@@ -165,7 +165,7 @@ var options: any = {
 app.listen(options, (err) => {
   if (err) throw err;
   const instanceId = process.env.NODE_APP_INSTANCE || '0';
-  console.info(`${moment().format('HH:mm:ss')} HIS-Connect API ${global.appDetail.version}-${global.appDetail.subVersion} started on port ${options.port}, PID: ${process.pid}`);
+  console.info(`${moment().format('HH:mm:ss')} HIS-Connect API ${global.appDetail.version}-${global.appDetail.subVersion} started on port ${options.port}, PID: ${process.pid} with NodeJS: ${process.version || ''}, Instance: ${instanceId}`);
 });
 
 // DB connection =========================================